reorganize code, move around lots of functions
