+++ jrnl/exception.py
@@ -1,0 +1,8 @@
+class UserAbort(Exception):
+    pass
+
+
+class UpgradeValidationException(Exception):
+    """Raised when the contents of an upgraded journal do not match the old journal"""
+
+    pass
