+++ jrnl/os_compat.py
@@ -1,0 +1,3 @@
+from sys import platform
+
+on_windows = "win32" in platform
